+++ eanalytics_api_py/earequest/datamining/payload.py
@@ -1,21 +1,21 @@
 product = {
-    "with-orderproduct" : 1,
-    "with-productparam" : 1,
-    "with-productgroup" : 1,
+    "with-orderproduct" : 1, #  purchased products
+    "with-productparam" : 1, # purchased products parameters
+    "with-productgroup" : 1, # purchased products groups
 }
 
 channel = {
-    "with-channel-count" : 1,
-    "with-channel-level" : 1,
-    "max-channel-info" : 9,
-    "max-channel-level" : 40,
-    "with-channel-via" : 1,
-    "with-channel-date" : 1,
-    "with-channel-profile" : 1,
-    "with-channel-mdevicetype" : 1,
-    "with-first-channel" : 1,
-    "with-last-channel" : 1,
-    "with-firstlast-channel-lag" : 1,
+    "with-channel-count" : 1, # number of marketing interactions
+    "with-channel-level" : 1, # with previous marketing interactions
+    "max-channel-info" : 14,  # marketing interactions detail level
+    "max-channel-level" : 40, # marketing interaction history length
+    "with-channel-via" : 1,   # attibution type pc/pw/r...
+    "with-channel-date" : 1,  # marketing interaction date
+    "with-channel-profile" : 1, # marketing interaction profile
+    "with-channel-mdevicetype" : 1, # marketing interaction device type
+    "with-first-channel" : 1, # with first marketing interaction
+    "with-last-channel" : 1, # with last marketing interaction
+    "with-firstlast-channel-lag" : 1, # delay (seconds) between first and last interaction
 }
 
 tpchannel = {
@@ -33,36 +33,36 @@ tpchannel = {
 }
 
 order = {
-    "with-ordertype" : 1,
-    "with-ordertypecustom" : 1,
-    "with-orderpayment" : 1,
-    "with-orderpage" : 1,
-    "with-orderpageurl" : 1,
-    "with-orderevent" : 1,
-    "with-scartdate" : 1,
-    "with-scartvalidmargin" : 1,
-    "with-realscartvalidmargin" : 1,
-    "with-scartvalidamount" : 1,
-    "with-realscartvalidamount" : 1,
+    "with-ordertype" : 1, # order type
+    "with-ordertypecustom" : 1, # order custom type
+    "with-orderpayment" : 1, # order payment
+    "with-orderpage" : 1, # path seen during the order session
+    "with-orderpageurl" : 1, # url seen during the order session
+    "with-orderevent" : 1, #  event triggered during the order session
+    "with-scartdate" : 1, # shopping cart date
+    "with-scartvalidmargin" : 1, # order margin
+    "with-realscartvalidmargin" : 1, # real order margin
+    "with-scartvalidamount" : 1, # order amount
+    "with-realscartvalidamount" : 1, # real order amount
 }
 
 analytics = {
     "with-extparam" : 1,
-    "with-geoloc" : 1,
-    "with-browser" : 1,
-    "with-lang" : 1,
-    "with-os" : 1,
-    "with-isp" : 1,
-    "with-action" : 1,
-    "with-mdevicetype" : 1,
-    "with-blockpromo" : 1,
-    "with-visitdate" : 1,
-    "with-visitlag" : 1,
-    "with-cgiparam" : 1,
-    "with-rebuyer-prevnb" : 1,
-    "with-rebuyer-prevamount" : 1,
-    "with-rebuyer-prevtm" : 1,
-    "with-rebuyer-prevtm-epoch" : 1,
+    "with-geoloc" : 1, # country Coded on 2 ISO-3166-1 characters
+    "with-browser" : 1, # browser name
+    "with-lang" : 1, # browser language
+    "with-os" : 1, # operating system
+    "with-isp" : 1, # internet service provider
+    "with-action" : 1, # action executed before purchasing
+    "with-mdevicetype" : 1, # type of device
+    "with-blockpromo" : 1, # last clicked auto promotion
+    "with-visitdate" : 1, # visit date
+    "with-visitlag" : 1, # duration of the visit
+    "with-cgiparam" : 1, # custome parameters
+    "with-rebuyer-prevnb" : 1, # previous number of orders
+    "with-rebuyer-prevamount" : 1, # previous revenue amount
+    "with-rebuyer-prevtm" : 1, # date of previous order
+    "with-rebuyer-prevtm-epoch" : 1, # date EPOCH of previous order
 }
 
 user = {
@@ -75,6 +75,6 @@ user = {
 }
 
 user_info = {
-    "with-cluster" : 1,
-    "with-iduserparam" :1,
+    "with-cluster" : 1, # Segment names
+    "with-iduserparam" :1, # CRM parameters
 }