+++ changelog.md
@@ -1,3 +1,10 @@
+### 0.0.95
+[extra documentation for datamining payload](eanalytics_api_py/earequest/datamining/payload.py)
+[new transformation module to work on pandas DataFrame](eanalytics_api_py/eatransform)
+[new functions in eajupyter](eanalytics_api_py/eajupyter/html.py)
+- button_hide_input_cells
+- remove_class_output_scroll
+
 ### 0.0.94
 internal optimizations